pyproject.toml
[tool.poetry]
name = "samokoder-core"
version = "0.1.0"
description = "Samokoder Core Backend"
authors = ["Samokoder Team"]
license = "FSL-1.1-MIT"
readme = "README.md"

repository = "https://github.com/samokoder-ai/samokoder-core"
homepage = "https://samokoder.ai"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/legacy/"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1,<5.0.0"
redis = "^5.0.1"
slowapi = "^0.1.9"
prometheus-client = "^0.19.0"
prometheus-fastapi-instrumentator = "^6.1.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"
arq = "^0.26.0"
openai = "^1.40.6"
anthropic = "^0.39.0"
groq = "^0.6.0"
jinja2 = "^3.1.4"
tiktoken = "^0.8.0"
sqlalchemy = { version = "^2.0.32", extras = ["asyncio"] }
aiosqlite = "^0.20.0"
psutil = "^5.9.8"
httpx = "^0.27.0"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
prompt-toolkit = "^3.0.48"
jsonref = "^1.1.0"
tenacity = "9.0.0"
docker = "^7.1.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0,<8.0.0"
ruff = "^0.3.5"
pytest-cov = "^5.0.0"
pytest-asyncio = ">=0.21.0,<0.24.0"
pytest-timeout = "^2.3.1"
pre-commit = "^3.7.0"

[tool.pytest.ini_options]
addopts = "-s --ignore=tests/integration --ignore=tests/old" # --cov=core --cov-report=xml
asyncio_mode = "auto"
python_paths = ["."]


[tool.coverage.report]
exclude_lines = ["if TYPE_CHECKING:", "raise NotImplementedError()"]
omit = ["core/db/migrations/*", "core/templates/tpl/*"]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py39"
lint.extend-select = ["I"]

[tool.bandit]
exclude_dirs = [
    "tests",
    ".venv",
    "venv",
    "build",
    "dist",
    "workspace",
]
skips = [
    "B101",  # assert_used (используем в тестах)
]
